brain/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aqa"
version = "0.5.0"
description = "Autonomous Quality Agent - LLM-powered API test generation and execution"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AQA Team", email = "aqa@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["testing", "api", "openapi", "llm", "automation", "quality"]
dependencies = [
    "pydantic>=2.0.0",
    "litellm>=1.0.0",
    "openapi-spec-validator>=0.7.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    # API dependencies
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
]

[project.urls]
Homepage = "https://github.com/lipeamarok/autonomous-quality-agent"
Repository = "https://github.com/lipeamarok/autonomous-quality-agent"
Documentation = "https://github.com/lipeamarok/autonomous-quality-agent#readme"

[project.scripts]
aqa = "src.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-requests>=0.41b0",
]

[tool.setuptools]
packages = ["src", "src.cli", "src.cli.commands", "src.ingestion", "src.generator", "src.adapter", "src.llm", "src.validator", "src.runner", "src.telemetry", "src.api", "src.api.routes", "src.api.schemas", "src.api.websocket"]
package-dir = {"" = "."}

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pyright]
include = ["src", "tests"]
exclude = [".venv", "__pycache__"]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic.*:",
]
markers = [
    "e2e: end-to-end tests that execute the real Runner binary",
    "slow: tests that take longer to run (network, compilation)",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E402",  # Module level import not at top of file (needed for circular import avoidance)
    "E501",  # Line too long (handled by formatter, not enforced strictly)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]  # Tests may have late imports
